Rename auth middleware to avoid name collision
