+++ backend/authMiddleware/auth.go
@@ -1,4 +1,4 @@
-package middleware
+package authMiddleware
 
 import (
 	"api/auth"
